Documentation: your checkout directory is "working tree" (part 2)
The same fix as the previous step, which replaced "work tree" with
"working tree" where appropriate, but this batch addresses the two
word phrase that are split across lines.

Since I didn't come up with a clever "grep" pattern to work across
line boundaries, I used pickaxe to find a block of "word 'work'
appears near the end of line, possibly followed by a run of
whitespaces or asterisks, then a newline, and then possibly a run of
whitespaces or asterisks, then word 'tree'" in the comparison
between an empty tree and what we have today, like so:

    $ LF=$'\n' WS="[* "$'\t'"]"
    $ git diff --pickaxe-regex -S"work$WS*$LF$WS*tree" \
	4b825dc642cb6 -- ':(glob)Documentation/*.adoc' |
      grep -A1 -e '^diff --git' -e "work${WS}$"

after finishing the previous two steps.

Signed-off-by: Junio C Hamano <gitster@pobox.com>