builtin/history: implement "reword" subcommand
Implement a new "reword" subcommand for git-history(1). This subcommand
is similar to the user performing an interactive rebase with a single
commit changed to use the "reword" instruction.

The major difference is that we do not check out the commit that is to
be reworded. This has the obvious benefit of being significantly faster
compared to git-rebase(1), but even more importantly it allows the user
to rewrite history even if there are local changes in the working tree
or in the index.

Signed-off-by: Patrick Steinhardt <ps@pks.im>
Signed-off-by: Junio C Hamano <gitster@pobox.com>